Address PR review findings across bonfire feature
P1 fixes:
- Fix G key crash in sidebar when no items (cursor = -1)
- Fix nondeterministic watermark rendering (break outside if)
- Redact llm_api_key in config set output
- Re-arm poll timers after Tab cycling in river view
- Include all rooms as fallback for fresh accounts with no bookmarks/recents
- Default Ollama model to llama3.2 when unspecified
- Validate cache_dir before layout save/load

P2 fixes:
- Add 30s timeout to LLM provider calls
- Use strings.Builder in prompt construction
- Cap error response reads at 1KB (OpenAI, Ollama)
- Use html.UnescapeString instead of manual entity decoding
- Fix cache comment (oldest-first, not LRU)
- Fix truncateRunes returning ellipsis when maxRunes <= 0
- Reset missCount in SetPollConfig for immediate effect
- Fix Color() overflow on 32-bit via uint32 modulo
- Fix variable shadowing in BonfireDigest goroutine
- Remove unused digestResult.idx field
- Capture awayStart before RecordTerminalFocus updates it
- Read-merge-write in MixerStore.Save
- Warn when llm_api_key silently dropped from local config
- Use plain text in FrontPage list titles (fix filtering)
- Clear RecordingID for chat URL routing
- Account for mixer height in river SetSize
- Call updatePollFocus on initial room load
- Guard empty tmux output in multiplexer
- Fix test: use os.Stat instead of filepath.Glob for existence check
- Increase TTL test margins for CI reliability

+++ internal/tui/workspace/data/types.go
@@ -270,13 +270,13 @@ func (r RoomID) Color(paletteSize int) int {
 	}
 	h := fnv.New32a()
 	h.Write([]byte(r.Key())) //nolint:errcheck // hash.Write never errors
-	return int(h.Sum32()) % paletteSize
+	return int(h.Sum32() % uint32(paletteSize))
 }
 
 // BonfireRoomConfig describes a discovered campfire room for bonfire.
 type BonfireRoomConfig struct {
 	RoomID
-	RoomName    string // project name or ping partner name
+	RoomName    string // campfire title (from the API's Title field)
 	ProjectName string
 }
